refactor: Refactor Midi trait to narrow the scope of each method and better encapsulate parsing concerns
`process_midi` receives a single MidiMessage and updates its state accordingly. This allows each Instrument to
effectively declare its capabilities with a simple match block. Expressing that state is the purview of a new
method `voice`. Refactored parsing code to promote clarity in light of a better understanding of USB MIDI Event
Packets.