feat: Make note priority user-configurable via pushbutton
A notable changes accompanies this feature:

All configuration moves up a level, out of the scope of the Micromoog and into the broader context of "some instrument." For most
configurations (e.g., note priority) this is probably fine. However, it does mean that all instruments will share the same set of
configuration possibilities, which may not make sense in all cases. For example, the InputMode configuration allows the user to
determine whether the note will be driven by the synth's KBD module or its OSC module, but some instruments (such as the Moog
Werkstatt-01) do not have a keyboard input at all. This choice moves the project forward now but may need to be reconsidered at
some point.