feat: Recalculate target note immediately upon changing note priority setting
For example, if, while holding C major chord, a user switches from low to high note priority, the
synth will immediately switch from playing the C to the A. Previously, the setting change would not
take effect until further MIDI input was received. That is, the user would have had to add or
remove a note in order to trigger the change.