refactor: Reorganize embargo state logic.
The reimplementation lets the Micromoog module decide and report for itself whether
to express incoming MIDI note events right away or to embargo them, and for how long.
(Unfortunately a considerable amount of logic is still required in main.rs to
interpret said report.) This is the core of the "chord cleanup" functionality. The
reorganization makes the voice_task easier to reason about (and extend). It is also
more efficient with regard to the number of tasks spawned.